ajout de la suppression de dossiers dans le gestionnaire de fichiers

+++ intranet/gestionnaire_fichier.php
@@ -1,6 +1,4 @@
 <!-- to do list :
-    - autoriser la suppression
-    - creer le dossier d'enregistrement si il n'existe pas
     - sanitariser les inputs utilisateurs
     - autoriser l'overwrite (si droits accordés)
     - sécuriser les fichiers de conf (empecher l'accès aux fichiers .json)
@@ -75,13 +73,14 @@ navigation();
     // echo $_SESSION["nom"];
     if (isset($_POST["add_file"])) {
         // echo "is working";
+        sanitise_post_inputs();
         $filetype = strtolower(pathinfo("gestionnaire_fichier".$_POST["path"].basename($_FILES["file-input"]["name"]),PATHINFO_EXTENSION));
         $filename = $_POST["name"]=="" ? $_FILES["file-input"]["name"] : $_POST["name"].".".$filetype;
         // echo "is working with name : $filename";
         if (path_exist("gestionnaire_fichier".$_POST["path"])) {
             if (isset($_SESSION["nom"]) && isset($_FILES["file-input"]["name"])) {
                 move_uploaded_file($_FILES["file-input"]["tmp_name"], "gestionnaire_fichier".$_POST["path"].$filename);
-                echo "gestionnaire_fichier".$_POST["path"].$filename;
+                // echo "gestionnaire_fichier".$_POST["path"].$filename;
                 $json = [
                     "owner" => $_SESSION["nom"],
                     "can_view" => [$_SESSION["nom"]],
@@ -91,10 +90,10 @@ navigation();
                 ];
 
                 file_put_contents("gestionnaire_fichier".$_POST["path"].$filename.".meta.json", json_encode($json));
-                // echo '<meta http-equiv="refresh" content="0;url=gestionnaire_fichier.php">';
+                echo '<meta http-equiv="refresh" content="0;url=gestionnaire_fichier.php">';
             } else {
                 echo "<script>alert('un probleme est survenu');</script>";
-                // echo '<meta http-equiv="refresh" content="0;url=gestionnaire_fichier.php">';
+                echo '<meta http-equiv="refresh" content="0;url=gestionnaire_fichier.php">';
             }
         }        
     }
@@ -103,7 +102,6 @@ navigation();
         $contenu = scandir($dossier);
         // c'est comme un `ls -l`, scandir retourne `.` et `..`, il faut les retirer
         $contenu = array_diff($contenu, [".", ".."]);
-
         foreach ($contenu as $index => $element) {
             $cheminComplet = $dossier."/".$element;
             $isLast = $index==array_key_last($contenu);
@@ -111,7 +109,14 @@ navigation();
 
             if (is_dir($cheminComplet) && user_can("view", $cheminComplet)) {
                 $nouveauPrefix = $prefix.($isLast ? "&nbsp&nbsp&nbsp&nbsp&nbsp" : "|&nbsp&nbsp&nbsp&nbsp");
-                echo $prefix.$branche."<strong>$element</strong><br>";
+                echo $prefix.$branche."<strong>$element</strong>";
+                if (user_can("delete", $cheminComplet)) {
+                    echo '<form action="" method="post" class="d-inline">
+                        <input type="text" name="filename" value="'.$cheminComplet.'" requiered hidden>
+                        <input type="submit" name="rm" value="icon à ajouter" class="bg-danger d-inline">
+                    </form>';
+                }
+                echo "<br>";
                 afficherArborescence($cheminComplet, $nouveauPrefix);
             }
             if (is_file($cheminComplet) && user_can("view", $cheminComplet)) {
@@ -121,11 +126,12 @@ navigation();
                 if (user_can("delete", $cheminComplet)) {
                     echo '<form action="" method="post" class="d-inline">
                         <input type="text" name="filename" value="'.$cheminComplet.'" requiered hidden>
-                        <input type="submit" name="rm_file" value="incon à ajouter" class="bg-danger d-inline">
+                        <input type="submit" name="rm" value="icon à ajouter" class="bg-danger d-inline">
                     </form>';
                 }
                 echo "<br>";
             }
+            
         }
     }
 
@@ -158,13 +164,18 @@ navigation();
         return $permissions["owner"]==$user;
     }
 
-    if (isset($_POST["rm_file"])) {
-        if (isset($_POST["filename"])) {
-            if (file_exists($_POST["filename"]) && user_can("delete", $_POST["filename"])) {
-                unlink($_POST["filename"]);
+    if (isset($_POST["rm"])) {
+        if (isset($_POST["filename"]) && user_can("delete", $_POST["filename"])) {
+            if (is_file($_POST["filename"])) {
+                try {
+                    unlink($_POST["filename"]);
+                } catch(Exception $e) {}
                 try {
                     unlink($_POST["filename"].".meta.json");
-                } catch(e) {}
+                } catch(Exception $e) {}
+                echo '<meta http-equiv="refresh" content="0;url=gestionnaire_fichier.php">';
+            } else if (is_dir($_POST["filename"])) {
+                rm_dir_r($_POST["filename"]);
                 echo '<meta http-equiv="refresh" content="0;url=gestionnaire_fichier.php">';
             }
         } else {
@@ -181,7 +192,7 @@ navigation();
             foreach (explode("/", $path) as $folder) {
                 $absoluteFolder .= "/".$folder;
                 $absoluteFolder = ltrim($absoluteFolder,"/");
-                echo $absoluteFolder." | ";
+                // echo $absoluteFolder." | ";
                 if (!is_dir($absoluteFolder)) create_folder($absoluteFolder);
                 // le chemin peut être crée ici mais pour une meilleure lisibilité on appelle une autre fonction
             }
@@ -192,7 +203,7 @@ navigation();
     function create_folder($path) {
         try {
             mkdir($path);
-            echo "successfuly created $path | ";
+            // echo "successfuly created $path | ";
             $json = [
                 "owner" => $_SESSION["nom"],
                 "can_view" => [$_SESSION["nom"]],
@@ -208,6 +219,26 @@ navigation();
         }
     }
 
+    function sanitise_post_inputs() {
+        if (isset($_POST["name"])) $_POST["name"] = htmlspecialchars(trim($_POST["name"]));
+        // pour le `path`, on retire les éventuels espaces, on s'assure de la présence de "/" au debut et a la fin du `path`
+        if (isset($_POST["path"])) $_POST["path"] = htmlspecialchars("/".trim(trim($_POST["path"],"/"))."/");
+        if (strpos($_POST["path"], "..")) echo "<script>alert('Par mesure de sécurité, la chaine \"..\" n\'est pas autorisée');</script>";
+    }
+
+    function rm_dir_r($dir) {
+        // echo "$dir | ";
+        if (!file_exists($dir)) return true;
+        if (!is_dir($dir)) return unlink($dir);
+        foreach (array_diff(scandir($dir), [".", ".."]) as $item) { // scan($dir) sans les "." et ".."
+            if (!rm_dir_r($dir."/".$item)) return false;
+        }
+        unlink("$dir.meta.json");
+        return rmdir($dir);
+    }
+
+?>
+<?php
     // Appel
     echo "<br>-------------------<br>";
     echo "gestionnaire_fichier<br>";
